move method to activate media file
